requirements.txt
evaluate==0.4.1
numpy==1.24.3
pandas==2.0.3
streamlit==1.30.0
torch==2.1.2+cu121
transformers==4.37.1
auto-gptq===0.6.0
optimum===1.16.2
python-docx===1.1.0